fix: VSCODE-63: syntax and snippets weren't included in .vsix
The `.vscodeignore` `src/*` means the entire src directory not included in packaged bundle.  So made scripts write these assets to resources instead. also having some fun making scripts pretty :p
