--- docs/README.md
@@ -1,14 +1,0 @@
-# design
-
-- using process developed with @rueckstiess 
-- design -> not writing gdocs etc.
-- instead start work -> document -> export stakeholder reports (fancy way of saying update lightweight JIRA tickets)
-
-## "groundrules"
-
-> (again, put the right work out first, report formats far second)
-
-- GitHub PR ready for review -> everything happens here
-- sketches, snippets, links, images, example/test scripts go in this directory :arrow_right: make pretty as design PR progresses
-- GitHub issues/projects/milestones are in the mix if they work 
-- TODO: lucas: something code of conduct like about attribution but cant find the words yet
