--- src/snippets/README.md
@@ -1,1 +1,0 @@
-Generated from mongodb-ace-autocompleter@^0.4.1
