--- src/syntaxes/README.md
@@ -1,7 +1,0 @@
-The `mongodb.tmLanguage.json` file is derived from [TypeScript.tmLanguage](https://github.com/Microsoft/TypeScript-TmLanguage/blob/master/TypeScript.tmLanguage) and injected with MongoDB symbols.
-
-To get the last version of the MongoDB grammar run the script:
-
-`npm run update-grammar`
-
-Check [vscode-mongodb-language](https://github.com/mongodb-js/vscode-mongodb-language) for more details.
